refactor: move authoring rule catalog out of model layer
RuleSeverity, RuleDefinition, _AUTHORING_RULE_DEFINITIONS, and catalog
API functions moved from model/constraints.py (1722->1204 lines) to
authoring/rule_catalog.py. RuleDefinition IDs, fix-hints, examples, and
docs are authoring/diagnostic metadata — runtime should not depend on them.

14 importers updated. Catalog placed at authoring/ level (not inside
rules/ subpackage) to avoid circular import through rules/__init__.py.
