+++ datamimic_ce/authoring/reference.py
@@ -43,14 +43,13 @@ from datamimic_ce.model.constraints import (
     Requires,
     RequiresWhenValue,
     ValidValues,
-    authoring_rule_definition,
-    authoring_rule_definitions,
     resolved_values,
     serialize_constraints,
-    serialize_rule_definition,
     serialize_source_capability,
     source_capabilities,
 )
+
+from datamimic_ce.authoring.rule_catalog import authoring_rule_definition, authoring_rule_definitions, serialize_rule_definition
 from datamimic_ce.model.element_registry import canonical_tag, element_aliases
 
 _GENERATOR_PACKAGE = "datamimic_ce.domains.common.literal_generators"
@@ -428,7 +427,6 @@ def capabilities_manifest() -> dict[str, Any]:
     from datamimic_ce.enums.converter_enums import ConverterEnum
     from datamimic_ce.enums.distribution_enums import POSITIONAL_NUMBER_SEQUENCES
     from datamimic_ce.exporters.exporter_util import buffered_exporter_names
-
     try:
         schema_version = version("datamimic_ce")
     except PackageNotFoundError:
